requirements.txt
torch>=1.10.0
pytorch-lightning==1.9.5
pytorch-forecasting==0.10.3
mlflow>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0